requirements.txt
sqlalchemy
PyMySQL
cryptography
selenium>=3.8.1
psutil>=5.4.6
youtube_dl>=2018.11.18
Pillow>=5.2.0
Requests>=2.20.0
PyYAML>=4.2b1
streamlink>=0.14.2
